requirements.txt
bcrypt==3.2.2
blinker==1.7.0
Flask>=2.0.1
Flask-JWT-Extended>=4.3.1
python-dotenv>=0.19.0
